requirements.txt
Pillow>=10.0.0
Faker>=24.0.0
numpy>=1.24.0
requests>=2.31.0
